pyproject.toml
[project]
name = "skills-monorepo-catalog"
version = "0.1.0"
description = "Tooling for a versioned skills monorepo catalog."
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0,<7.0",
]

[tool.setuptools]
packages = ["skill_catalog"]
